requirements.txt
allure-pytest==2.9.45
Faker==13.3.2
pytest==7.0.1
requests==2.27.1
